+++ MovieDbLite.MVC/Models/MovieCrewMember.cs
@@ -4,7 +4,7 @@
     {
         public long MovieId { get; set; }
         public long FilmMemberId { get; set; }
-        public int FilmRoleId { get; set; }
+        public short FilmRoleId { get; set; }
 
         public virtual FilmMember FilmMember { get; set; }
         public virtual FilmRole FilmRole { get; set; }
