+++ MovieDbLite.MVC/Models/MovieGenre.cs
@@ -3,7 +3,7 @@
     public partial class MovieGenre
     {
         public long MovieId { get; set; }
-        public int GenreId { get; set; }
+        public short GenreId { get; set; }
 
         public virtual Genre Genre { get; set; }
         public virtual Movie Movie { get; set; }
